requirements.txt
Flask
Flask-DebugToolbar
Flask-WTF
WTForms
cryptography
pyaudio
opencv-python
requests
pinatapy
numpy
scipy
scikit-learn
python_speech_features
IPython
retry
py-solc-x
web3
pandas
matplotlib
seaborn
gunicorn
